persistence: fix every actionable finding from the three 613229c reviews
All claims rechecked against raw objdump on loco.exe (Ghidra/Fabric
unavailable); every required fix landed, two rejected with raw proof.

- INPUT_FindObjectAt default mode (0x41E498): the pick range is now the
  16-bit +0x158 count of the typed g_resmgr.GetById(mode) lookup
  (0x446EA0), NOT a collection scan (0x41E4AD/0x41E4CA); host returns
  nullptr for non-positive lookups (the binary checks only ==0 and would
  read 0x157 on a -1 error; documented hardening).
- INPUT_LoadSaveFile placement deltas use floor division (floor_div2 =
  the cltd/sub/sar idiom at 0x41D6A9..0x41D6D2): negative odd deltas
  divide like the x86 (-3/2 = -2), not C++ truncation.
- SaveCurrentWorld level-table name owner corrected to *(obj+0x10)
  (sub-slot[0]; 0x41DC78 mov (%ebx),%edi with ebx=obj+0x10) + null guard.
- ResDataSave: host record buffers moved to native typed RESDATA members
  (host_record_entity/vehicle under UINTPTR_MAX != 32); the x86 +0x04/
  +0x84 record offsets would land inside pointer-width host members.
- _WIN32 write path now uses the 0x465090 WRITE-stream ctor (mode|2,
  WIN32_StreamOpenWriteFile) instead of the read ctor 0x463970 (mode|1);
  the old "CRT_floor/ceil" labels for 0x465090/0x465010 corrected.
- Distinct canonical globals: g_allow_building_placement = 0x4FD3DC
  (loader/placement flag) vs g_is_town_mode = 0x485328 (tilemap/town
  flag); tilemap.h/tilemap.cpp/link_stubs/fixtures split the conflated
  symbol; the 4 town-mode tilemap gates remapped by binary address.
- Atomic host save: HostSaveStream stages "<path>.tmp" + Commit() renames
  (host_save_commit); uncommitted temps removed on destruction;
  PersistenceAdapter::write_save now implements its documented temp+rename
  contract; SaveCurrentWorld checks every write, commits, and returns 0 on
  any failure — no partial curr ever appears.
- seed_fresh_world_from_fixture propagates INPUT_SaveCurrentWorld failure
  (returns false; HostLoadingSequence logs loudly) — a fresh seed never
  reports success without a durable curr.
- SaveCurrentWorld gets the symmetric name_escapes guard; strict 16 MiB
  preview cap in the host load/write/save-preview paths; NUL-safe bounded
  "PARTY" child-name probe; NewWorld docs corrected (the binary divides
  the re-read entity_count by 10, 0x41E181..0x41E1D0, so the animated
  branch is constant per loop); LoadWorld's ".sav" companion no longer
  overwrites the host current-name bookkeeping (snapshot/restore).
- RESDATA_IsBuildingTile canonicalised as uint8_t __fastcall(int32_t) in
  tilemap.h with a real 0x44BD30 implementation in tilemap.cpp; the old
  InputMgr.cpp plain `int(intptr_t)` extern bound to a null symbol under
  the ignore-all link.
- DirToOffset ABI preserved exactly: the four helpers return the OUTPUT
  POINTER (binary EAX) and Netman.cpp dereferences
  (`off = *INPUT_DirToOffset_Left(&off)`), matching the 0x43E252
  mov (%eax),%ecx call sites instead of the decompiler-intent value ABI.
- Host event-stream fix (pre-existing flake): the SDL pump's raw ESC
  fallback IS the mode-10 quit path but never emitted mode_changed(10),
  so test_main_menu_escape_exits_cleanly flaked ~50% whenever the Sway
  content-rect y-offset race (25px vs 0px) missed the name-field click;
  the fallback now emits mode_changed(old,10) (host-only).

Tests: input_world_test gains FindObjectAt default-mode range (MAP_32BIT
fake resource) + not-found paths, read-only-dir save failure, failed
atomic rename (no partial/temp), curr.sav bookkeeping, symmetric escape
guard, seed success and seed-failure-without-durable-curr; persistence_
adapter_test gains atomic-write no-temp and unwritable-dir checks.

Rejected with raw proof: (a) NewWorld's scroll-loop code was already
right — the disassembly divides entity_count by 10, not the index, so
only the docs were wrong (fixed); (b) the negative-odd placement deltas
are not observable through the public persistence API (placement is gated
behind host_placement_available) — the fix is in place and documented.

make (135/135 objects), make check 135/135, make test 26 PASS,
make test-integration 11/11 (twice consecutively) in nix develop.

+++ src/decompiled_cpp/shared/link_stubs.cpp
@@ -501,7 +501,7 @@ void* _g_train = nullptr; void* DAT_004a97a0 = nullptr; void* DAT_004a9994 = nul
 void* g_click_on_town = nullptr; void* g_click_on_building = nullptr; void* g_cgwnd = nullptr;
 int32_t g_clean_exit = 0; void* g_client_rect = nullptr; void* _g_cursor_surface = nullptr;
 void* g_frame_event = nullptr; void* g_fullscreen_rect = nullptr; void* g_game_config = nullptr;
-int32_t g_game_difficulty = 0; int32_t g_in_build_mode = 0; int32_t g_is_town_mode = 0;
+int32_t g_game_difficulty = 0; int32_t g_in_build_mode = 0; uint8_t g_is_town_mode = 0;
 void* g_nameEntryPanel = nullptr; void* g_netSettings = nullptr; void* _g_network_queue = nullptr;
 void* g_network_queue = nullptr; void* _g_network_thread = nullptr; void* g_network_thread = nullptr;
 void* g_pixel_data_cache = nullptr; int32_t g_placement_resource_id = 0; int32_t g_ref_count = 0;
